+++ SourceryTests/Parsing/AnnotationsParserSpec.swift
@@ -104,6 +104,26 @@ class AnnotationsParserSpec: QuickSpec {
 
                     expect(result["decoding"] as? Annotations).to(equal(annotations))
                 }
+
+                it("extracts json string annotations into array") {
+                    let parsedAnnotations = parse("// sourcery: theArray=\"[22,55,88]\"")
+                    expect(parsedAnnotations["theArray"] as? [Int]).to(equal([22, 55, 88]))
+                }
+
+                it("extracts json string annotations into arrays of dictionaries") {
+                    let parsedAnnotations = parse("// sourcery: propertyMapping=\"[{\"from\": \"lockVersion\", \"to\": \"version\"},{\"from\": \"goalStatus\", \"to\": \"status\"}]\"")
+                    expect(parsedAnnotations["propertyMapping"] as? [[String: String]]).to(equal([["from": "lockVersion", "to": "version"], ["from": "goalStatus", "to": "status"]]))
+                }
+
+                it("extracts json string annotations into dictionary") {
+                    let parsedAnnotations = parse("// sourcery: theDictionary=\"{\"firstValue\": 22,\"secondValue\": 55}\"")
+                    expect(parsedAnnotations["theDictionary"] as? [String: Int]).to(equal(["firstValue": 22, "secondValue": 55]))
+                }
+
+                it("extracts json string annotations into dictionaries of arrays") {
+                    let parsedAnnotations = parse("// sourcery: theArrays=\"{\"firstArray\":[22,55,88],\"secondArray\":[1,2,3,4]}\"")
+                    expect(parsedAnnotations["theArrays"] as? [String: [Int]]).to(equal(["firstArray": [22, 55, 88], "secondArray": [1, 2, 3, 4]]))
+                }
             }
         }
     }