+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 # Sourcery CHANGELOG
 
 ---
+## Master
+
+### New Features
+
+- You can now pass a json string as a command line arg or annotation and have it parsed into a Dictionary or Array to be used in the template.
+
 
 ## Master